Move server into services
